backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.21",
    "fastapi>=0.116.1",
    "numpy>=2.3.2",
    "openai-agents>=0.2.4",
    "pandas>=2.3.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "urllib3>=2.5.0",
    "uvicorn>=0.35.0",
]
